+++ lib/nebulex/distributed/application.ex
@@ -3,23 +3,17 @@ defmodule Nebulex.Distributed.Application do
 
   use Application
 
+  alias Nebulex.Distributed.Cluster
+
   @impl true
   def start(_type, _args) do
-    children = pg_children()
+    children = [
+      Cluster.pg_child_spec()
+    ]
 
     Supervisor.start_link(children,
       strategy: :one_for_one,
       name: Nebulex.Distributed.Supervisor
     )
-  end
-
-  if Code.ensure_loaded?(:pg) do
-    defp pg_children do
-      [%{id: :pg, start: {:pg, :start_link, [Nebulex.Distributed.Cluster]}}]
-    end
-  else
-    defp pg_children do
-      []
-    end
   end
 end